refactor: fold query error seam types into factory module
